Refactor skills panel and fix large file finder
